Fix and optimize error handling in console execution and request handler context
