+++ build.default.properties
@@ -8,7 +8,7 @@
 ;--------------------------------------------------------------------------------
 
 ; ---- Module Release Settings --------------------------------------------------
-release.version = 1.1.5-beta5
+release.version = 1.1.5-beta6
 release.name = Iron Knight
 
 ; ---- PHPCPD Settings ----------------------------------------------------------
